+++ nautobot_plugin_nornir/contrib/constants.py
@@ -1,0 +1,15 @@
+"""Constants used by other contrib utilities."""
+
+FIELDS_PK = {
+    "platform",
+    "tenant_group",
+    "tenant",
+    "location",
+    "role",
+    "rack",
+    "rack_group",
+    "manufacturer",
+    "device_type",
+}
+
+FIELDS_NAME = {"tags", "status"}
