+++ nautobot_plugin_nornir/jobs.py
@@ -3,17 +3,21 @@
 from datetime import datetime
 
 import yaml
-from nautobot.apps.jobs import BooleanVar, Job, MultiObjectVar, register_jobs
-from nautobot.dcim.filters import DeviceFilterSet
-from nautobot.dcim.models import Device, DeviceType, Location, Manufacturer, Platform, Rack, RackGroup
-from nautobot.extras.models import Role, Status, Tag
-from nautobot.tenancy.models import Tenant, TenantGroup
+from nautobot.apps.jobs import MultiObjectVar, register_jobs
+
+# from nautobot.dcim.filters import DeviceFilterSet
+from nautobot.dcim.models import Device  # , DeviceType, Location, Manufacturer, Platform, Rack, RackGroup
+
+# from nautobot.extras.models import Role, Status, Tag
+# from nautobot.tenancy.models import Tenant, TenantGroup
 from nornir import InitNornir
 from nornir.core.plugins.inventory import InventoryPluginRegister
-from nornir_nautobot.exceptions import NornirNautobotException
+
+# from nornir_nautobot.exceptions import NornirNautobotException
 from nornir_nautobot.plugins.tasks.dispatcher import dispatcher
 
-from nautobot_plugin_nornir.constants import FIELDS_NAME, FIELDS_PK, NORNIR_SETTINGS
+from nautobot_plugin_nornir.constants import NORNIR_SETTINGS
+from nautobot_plugin_nornir.contrib.jobs import NornirBaseJob
 from nautobot_plugin_nornir.logger import NornirLogger
 from nautobot_plugin_nornir.plugins.inventory.nautobot_orm import NautobotORMInventory
 from nautobot_plugin_nornir.plugins.processors.processor import ProcessNornirResults
@@ -21,65 +25,7 @@ from nautobot_plugin_nornir.plugins.processors.processor import ProcessNornirRes
 InventoryPluginRegister.register("nautobot-inventory", NautobotORMInventory)
 
 
-def get_job_filter(data=None):
-    """Helper function to return a the filterable list of OS's based on platform.name and a specific custom value."""
-    if not data:
-        data = {}
-    query = {}
-
-    # Translate instances from FIELDS set to list of primary keys
-    for field in FIELDS_PK:
-        if data.get(field):
-            query[field] = data[field].values_list("pk", flat=True)
-
-    # Translate instances from FIELDS set to list of names
-    for field in FIELDS_NAME:
-        if data.get(field):
-            query[field] = data[field].values_list("name", flat=True)
-
-    if data.get("device") and isinstance(data["device"], Device):
-        query.update({"id": [str(data["device"].pk)]})
-    elif data.get("device"):
-        query.update({"id": data["device"].values_list("pk", flat=True)})
-
-    base_qs = Device.objects.filter()
-    devices_filtered = DeviceFilterSet(data=query, queryset=base_qs)
-    devices_no_platform = devices_filtered.qs.filter(platform__isnull=True)
-    if devices_no_platform.exists():
-        raise NornirNautobotException(
-            f"`E3017:` The following device(s) {', '.join([device.name for device in devices_no_platform])} have no platform defined. Platform is required."
-        )
-
-    return devices_filtered.qs
-
-
-class FormEntry:  # pylint: disable=too-few-public-methods
-    """Class definition to use as Mixin for form definitions."""
-
-    tenant_group = MultiObjectVar(model=TenantGroup, required=False)
-    tenant = MultiObjectVar(model=Tenant, required=False)
-    location = MultiObjectVar(model=Location, required=False)
-    rack_group = MultiObjectVar(model=RackGroup, required=False)
-    rack = MultiObjectVar(model=Rack, required=False)
-    role = MultiObjectVar(model=Role, required=False)
-    manufacturer = MultiObjectVar(model=Manufacturer, required=False)
-    platform = MultiObjectVar(model=Platform, required=False)
-    device_type = MultiObjectVar(model=DeviceType, required=False, display_field="display_name")
-    device = MultiObjectVar(model=Device, required=False)
-    tags = MultiObjectVar(
-        model=Tag, required=False, display_field="name", query_params={"content_types": "dcim.device"}
-    )
-    status = MultiObjectVar(
-        model=Status,
-        required=False,
-        query_params={"content_types": Device._meta.label_lower},
-        display_field="label",
-        label="Device Status",
-    )
-    debug = BooleanVar(description="Enable for more verbose debug logging")
-
-
-class DebugInventoryJob(Job):
+class DebugInventoryJob(NornirBaseJob):
     """Job to debug Nornir Inventory."""
 
     # Device only instead of FormEntry to limit scope as this can be very verbose.
@@ -96,7 +42,7 @@ class DebugInventoryJob(Job):
     def run(self, *args, **kwargs):  # pylint: disable=too-many-branches
         """Run config compliance report script."""
         now = datetime.now()
-        qs = get_job_filter(data=kwargs)
+        qs = self.get_device_queryset(data=kwargs)
         after_qs = datetime.now()
         self.logger.info("Queryset Took\n%s", after_qs - now)
         self.logger.info("Queryset Count\n%s", qs.count())
@@ -139,7 +85,7 @@ class DebugInventoryJob(Job):
             raise
 
 
-class ConnectivityCheckTask(FormEntry, Job):  # pylint: disable=too-many-instance-attributes
+class ConnectivityCheckTask(NornirBaseJob):  # pylint: disable=too-many-instance-attributes
     """Nautobot Job for validating connectivity and authentication."""
 
     class Meta:  # pylint: disable=too-few-public-methods
@@ -153,7 +99,7 @@ class ConnectivityCheckTask(FormEntry, Job):  # pylint: disable=too-many-instanc
     def run(self, *args, **kwargs):
         """Run connectivity check and authentication tests."""
         logger = NornirLogger(self.job_result, self.logger.getEffectiveLevel())
-        qs = get_job_filter(data=kwargs)
+        qs = self.get_device_queryset(data=kwargs)
         try:
             with InitNornir(
                 runner=NORNIR_SETTINGS.get("runner"),
